+++ CHANGELOG.md
@@ -22,6 +22,10 @@
 
   Bolt will now lookup the default PuppetDB config at `%COMMON_APPDATA%\PuppetLabs\client-tools\puppetdb.conf` instead of the hardcoded path `C:\ProgramData\PuppetLabs\client-tools\puppetdb.conf`.
 
+* **Bolt could not find plans in subdirectories of `plans` directory**
+
+  Bolt will now correctly search for subdir paths under the `plans` directory for plan names when determining if the plan is a Puppet plan or yaml plan.
+
 ## Bolt 1.41.0
 
 ### New features